# artifacts/api-server/src/modules/viewer/viewer.controller.ts
import {
  Body,
  Controller,
  Get,
  HttpStatus,
  Inject,
  Logger,
  Param,
  Patch,
  Post,
  Req,
  Res,
} from "@nestjs/common";
import { ApiBearerAuth, ApiOperation, ApiTags } from "@nestjs/swagger";
import { z, ZodError } from "zod";
import type { Request, Response } from "express";
import { AuthService } from "../auth/auth.service";
import { DiscoveryService } from "../discovery/discovery.service";
import { GoalsService } from "../goals/goals.service";
import { ViewerService } from "./viewer.service";
import { API_TAGS } from "../../docs/openapi/tags";
import { MATCHA_BEARER_AUTH } from "../../docs/openapi/security";

const profileUpdateSchema = z.object({
  name: z.string().trim().min(2).max(120).optional(),
  dateOfBirth: z.string().regex(/^\d{4}-\d{2}-\d{2}$/).optional(),
  location: z.string().trim().max(255).optional(),
  country: z.string().trim().max(120).optional(),
  profession: z.string().trim().max(120).optional(),
  genderIdentity: z.string().trim().max(64).optional(),
  pronouns: z.string().trim().max(64).optional(),
  personality: z.string().trim().max(64).optional(),
  relationshipGoals: z.string().trim().max(120).optional(),
  languagesSpoken: z.array(z.string().trim().min(1).max(64)).max(7).optional(),
  education: z.string().trim().max(120).optional(),
  childrenPreference: z.string().trim().max(120).optional(),
  physicalActivity: z.string().trim().max(120).optional(),
  alcoholUse: z.string().trim().max(120).optional(),
  tobaccoUse: z.string().trim().max(120).optional(),
  politicalInterest: z.string().trim().max(120).optional(),
  religionImportance: z.string().trim().max(120).optional(),
  religion: z.string().trim().max(120).optional(),
  bio: z.string().trim().max(2000).optional(),
  bodyType: z.string().trim().max(120).optional(),
  height: z.string().trim().max(32).optional(),
  hairColor: z.string().trim().max(120).optional(),
  ethnicity: z.string().trim().max(160).optional(),
  interests: z.array(z.string().trim().min(1).max(64)).max(24).optional(),
  latitude: z.number().min(-90).max(90).optional(),
  longitude: z.number().min(-180).max(180).optional(),
});

const settingsUpdateSchema = z.object({
  language: z.enum(["es", "en"]).optional(),
  heightUnit: z.enum(["metric", "imperial"]).optional(),
  genderIdentity: z.string().trim().max(64).optional(),
  pronouns: z.string().trim().max(64).optional(),
  personality: z.string().trim().max(64).optional(),
});

const goalCompletionSchema = z.object({
  completed: z.literal(true).optional().default(true),
});

const goalReorderSchema = z.object({
  category: z.enum([
    "physical",
    "personality",
    "family",
    "expectations",
    "language",
    "studies",
  ]),
  orderedGoalKeys: z.array(z.string().trim().min(1).max(64)).min(1),
});

@ApiTags(API_TAGS.viewer)
@ApiBearerAuth(MATCHA_BEARER_AUTH)
@Controller()
export class ViewerController {
  private readonly logger = new Logger(ViewerController.name);

  constructor(
    @Inject(AuthService) private readonly authService: AuthService,
    @Inject(ViewerService) private readonly viewerService: ViewerService,
    @Inject(GoalsService) private readonly goalsService: GoalsService,
    @Inject(DiscoveryService) private readonly discoveryService: DiscoveryService
  ) {}

  private getAuthorizationHeader(req: Request) {
    return typeof req.headers.authorization === "string"
      ? req.headers.authorization
      : undefined;
  }

  private getRequestId(req: Request) {
    const header = req.headers["x-matcha-request-id"];
    if (typeof header === "string" && header.trim()) {
      return header.trim();
    }
    if (Array.isArray(header) && header[0]?.trim()) {
      return header[0].trim();
    }
    return undefined;
  }

  private getLocationSource(req: Request) {
    const header = req.headers["x-matcha-location-source"];
    if (typeof header === "string" && header.trim()) {
      return header.trim().slice(0, 64);
    }
    if (Array.isArray(header) && header[0]?.trim()) {
      return header[0].trim().slice(0, 64);
    }
    return undefined;
  }

  private sendAuthError(res: Response, error: unknown) {
    const message =
      error instanceof Error && error.message ? error.message : "UNAUTHORIZED";
    return res.status(HttpStatus.UNAUTHORIZED).json({ error: message });
  }

  @Get("viewer/bootstrap")
  @ApiOperation({ summary: "Get bootstrap data for the authenticated viewer shell" })
  async getViewerBootstrap(@Req() req: Request, @Res() res: Response) {
    try {
      const auth = await this.authService.authenticate(this.getAuthorizationHeader(req), {
        requestId: this.getRequestId(req),
        source: "viewer_bootstrap",
      });
      return res.json(await this.viewerService.getBootstrap(auth.user.id));
    } catch (error) {
      return this.sendAuthError(res, error);
    }
  }

  @Get("me/profile")
  @ApiOperation({ summary: "Get the full editable viewer profile" })
  async getProfile(@Req() req: Request, @Res() res: Response) {
    try {
      const auth = await this.authService.authenticate(this.getAuthorizationHeader(req), {
        requestId: this.getRequestId(req),
        source: "viewer_get_profile",
      });
      return res.json(await this.viewerService.getProfile(auth.user.id));
    } catch (error) {
      return this.sendAuthError(res, error);
    }
  }

  @Patch("me/profile")
  @ApiOperation({ summary: "Update the full viewer profile" })
  async updateProfile(@Req() req: Request, @Body() body: unknown, @Res() res: Response) {
    try {
      const auth = await this.authService.authenticate(this.getAuthorizationHeader(req), {
        requestId: this.getRequestId(req),
        source: "viewer_update_profile",
      });
      const payload = profileUpdateSchema.parse(body);
      const requestId = this.getRequestId(req);
      this.logger.log(
        `[viewer-profile-update] ${JSON.stringify({
          requestId: requestId || null,
          userId: auth.user.id,
          fields: Object.keys(payload || {}),
          locationSource: this.getLocationSource(req) || null,
        })}`
      );
      return res.json(
        await this.viewerService.updateProfile(auth.user.id, payload, {
          requestId,
          locationSource: this.getLocationSource(req),
        })
      );
    } catch (error) {
      if (error instanceof ZodError) {
        return res.status(HttpStatus.BAD_REQUEST).json({
          error: "INVALID_PROFILE_PAYLOAD",
          issues: error.flatten(),
        });
      }
      if (error instanceof Error && error.message) {
        if (error.message === "UNDERAGE" || error.message === "INVALID_DATE_OF_BIRTH") {
          return res.status(HttpStatus.BAD_REQUEST).json({ error: error.message });
        }
        return this.sendAuthError(res, error);
      }
      console.error(error);
      return res
        .status(HttpStatus.INTERNAL_SERVER_ERROR)
        .json({ error: "INTERNAL_SERVER_ERROR" });
    }
  }

  @Get("me/settings")
  @ApiOperation({ summary: "Get viewer settings" })
  async getSettings(@Req() req: Request, @Res() res: Response) {
    try {
      return res.json(
        await this.authService.getSettings(this.getAuthorizationHeader(req))
      );
    } catch (error) {
      return this.sendAuthError(res, error);
    }
  }

  @Patch("me/settings")
  @ApiOperation({ summary: "Update viewer settings" })
  async updateSettings(@Req() req: Request, @Body() body: unknown, @Res() res: Response) {
    try {
      const payload = settingsUpdateSchema.parse(body);
      return res.json(
        await this.authService.updateSettings(this.getAuthorizationHeader(req), payload)
      );
    } catch (error) {
      if (error instanceof ZodError) {
        return res.status(HttpStatus.BAD_REQUEST).json({
          error: "INVALID_SETTINGS_PAYLOAD",
          issues: error.flatten(),
        });
      }
      if (error instanceof Error && error.message) {
        return this.sendAuthError(res, error);
      }
      console.error(error);
      return res
        .status(HttpStatus.INTERNAL_SERVER_ERROR)
        .json({ error: "INTERNAL_SERVER_ERROR" });
    }
  }

  @Get("me/goals")
  @ApiOperation({ summary: "List the current viewer goal tasks" })
  async getGoals(@Req() req: Request, @Res() res: Response) {
    try {
      const auth = await this.authService.authenticate(this.getAuthorizationHeader(req));
      return res.json(await this.goalsService.getUserGoals(auth.user.id));
    } catch (error) {
      return this.sendAuthError(res, error);
    }
  }

  @Patch("me/goals/:goalKey")
  @ApiOperation({ summary: "Mark a viewer goal task as completed" })
  async completeGoal(
    @Req() req: Request,
    @Param("goalKey") goalKey: string,
    @Body() body: unknown,
    @Res() res: Response
  ) {
    try {
      const auth = await this.authService.authenticate(this.getAuthorizationHeader(req));
      goalCompletionSchema.parse(body || {});
      return res.json(await this.goalsService.completeGoalTask(auth.user.id, goalKey));
    } catch (error) {
      if (error instanceof ZodError) {
        return res.status(HttpStatus.BAD_REQUEST).json({
          error: "INVALID_GOAL_UPDATE_PAYLOAD",
          issues: error.flatten(),
        });
      }
      if (error instanceof Error && error.message) {
        if (error.message === "GOAL_NOT_FOUND") {
          return res.status(HttpStatus.NOT_FOUND).json({ error: error.message });
        }
        return this.sendAuthError(res, error);
      }
      console.error(error);
      return res
        .status(HttpStatus.INTERNAL_SERVER_ERROR)
        .json({ error: "INTERNAL_SERVER_ERROR" });
    }
  }

  @Post("me/goals/reorder")
  @ApiOperation({ summary: "Reorder the viewer goal tasks within one category" })
  async reorderGoals(@Req() req: Request, @Body() body: unknown, @Res() res: Response) {
    try {
      const auth = await this.authService.authenticate(this.getAuthorizationHeader(req));
      const payload = goalReorderSchema.parse(body);
      return res.json(
        await this.goalsService.reorderGoalTasks(
          auth.user.id,
          payload.category,
          payload.orderedGoalKeys
        )
      );
    } catch (error) {
      if (error instanceof ZodError) {
        return res.status(HttpStatus.BAD_REQUEST).json({
          error: "INVALID_GOAL_REORDER_PAYLOAD",
          issues: error.flatten(),
        });
      }
      if (error instanceof Error && error.message) {
        if (error.message === "INVALID_GOAL_ORDER") {
          return res.status(HttpStatus.BAD_REQUEST).json({ error: error.message });
        }
        return this.sendAuthError(res, error);
      }
      console.error(error);
      return res
        .status(HttpStatus.INTERNAL_SERVER_ERROR)
        .json({ error: "INTERNAL_SERVER_ERROR" });
    }
  }

  @Post("me/goals/unlock/seen")
  @ApiOperation({ summary: "Mark the goals unlock notice as seen" })
  async markGoalsUnlockSeen(@Req() req: Request, @Res() res: Response) {
    try {
      const auth = await this.authService.authenticate(this.getAuthorizationHeader(req));
      return res.json(await this.goalsService.markGoalsUnlockSeen(auth.user.id));
    } catch (error) {
      return this.sendAuthError(res, error);
    }
  }

  @Get("me/discovery/preferences")
  @ApiOperation({ summary: "Get discovery preferences from the viewer surface" })
  async getDiscoveryPreferences(@Req() req: Request, @Res() res: Response) {
    try {
      const auth = await this.authService.authenticate(this.getAuthorizationHeader(req));
      return res.json(await this.discoveryService.getPreferences(auth.user.id));
    } catch (error) {
      return this.sendAuthError(res, error);
    }
  }

  @Patch("me/discovery/preferences")
  @ApiOperation({ summary: "Update discovery preferences from the viewer surface" })
  async updateDiscoveryPreferences(
    @Req() req: Request,
    @Body() body: unknown,
    @Res() res: Response
  ) {
    try {
      const auth = await this.authService.authenticate(this.getAuthorizationHeader(req));
      const payload = z
        .object({
          filters: z
            .object({
              selectedGenders: z
                .array(z.enum(["male", "female", "non_binary", "fluid"]))
                .max(4),
              therianMode: z.enum(["exclude", "include", "only"]),
              ageMin: z.number().int().min(18).max(100),
              ageMax: z.number().int().min(18).max(100),
            })
            .refine((value) => value.ageMin <= value.ageMax),
        })
        .parse(body);
      this.logger.log(
        `[viewer-discovery-preferences] ${JSON.stringify({
          requestId: this.getRequestId(req) || null,
          userId: auth.user.id,
          filters: payload.filters,
        })}`
      );
      return res.json(
        await this.discoveryService.updatePreferences(auth.user.id, payload.filters)
      );
    } catch (error) {
      if (error instanceof ZodError) {
        return res.status(HttpStatus.BAD_REQUEST).json({
          error: "INVALID_DISCOVERY_PREFERENCES_PAYLOAD",
          issues: error.flatten(),
        });
      }
      if (error instanceof Error && error.message) {
        return this.sendAuthError(res, error);
      }
      console.error(error);
      return res
        .status(HttpStatus.INTERNAL_SERVER_ERROR)
        .json({ error: "INTERNAL_SERVER_ERROR" });
    }
  }
}