# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/viewer/bootstrap": {
      "get": {
        "operationId": "Viewer_getViewerBootstrap",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Get bootstrap data for the authenticated viewer shell",
        "tags": [
          "viewer"
        ]
      }
    },
    "/api/me/profile": {
      "get": {
        "operationId": "Viewer_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Get the full editable viewer profile",
        "tags": [
          "viewer"
        ]
      },
      "patch": {
        "operationId": "Viewer_updateProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Update the full viewer profile",
        "tags": [
          "viewer"
        ]
      }
    },
    "/api/me/settings": {
      "get": {
        "operationId": "Viewer_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Get viewer settings",
        "tags": [
          "viewer"
        ]
      },
      "patch": {
        "operationId": "Viewer_updateSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Update viewer settings",
        "tags": [
          "viewer"
        ]
      }
    },
    "/api/me/goals": {
      "get": {
        "operationId": "Viewer_getGoals",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "List the current viewer goal tasks",
        "tags": [
          "viewer"
        ]
      }
    },
    "/api/me/goals/{goalKey}": {
      "patch": {
        "operationId": "Viewer_completeGoal",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Mark a viewer goal task as completed",
        "tags": [
          "viewer"
        ]
      }
    },
    "/api/me/goals/reorder": {
      "post": {
        "operationId": "Viewer_reorderGoals",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Reorder the viewer goal tasks within one category",
        "tags": [
          "viewer"
        ]
      }
    },
    "/api/me/goals/unlock/seen": {
      "post": {
        "operationId": "Viewer_markGoalsUnlockSeen",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Mark the goals unlock notice as seen",
        "tags": [
          "viewer"
        ]
      }
    },
    "/api/me/discovery/preferences": {
      "get": {
        "operationId": "Viewer_getDiscoveryPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Get discovery preferences from the viewer surface",
        "tags": [
          "viewer"
        ]
      },
      "patch": {
        "operationId": "Viewer_updateDiscoveryPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Update discovery preferences from the viewer surface",
        "tags": [
          "viewer"
        ]
      }
    }
  },
  "info": {
    "title": "Matcha API",
    "description": "Canonical API contract for the Matcha backend. Target OpenAPI 3.1 where the active toolchain remains smooth.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "matcha-bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Bearer access token returned by Matcha authentication endpoints."
      }
    },
...
  }
}